LC-949 출시알림버튼 파일 분리
